+++ src/main/groovy/ru/vyarus/gradle/plugin/quality/QualityPlugin.groovy
@@ -10,6 +10,7 @@ import org.gradle.api.execution.TaskExecutionGraph
 import org.gradle.api.plugins.GroovyPlugin
 import org.gradle.api.plugins.JavaPlugin
 import org.gradle.api.plugins.quality.*
+import org.gradle.api.reporting.Report
 import org.gradle.api.tasks.SourceSet
 import org.gradle.api.tasks.SourceTask
 import org.gradle.api.tasks.TaskProvider
@@ -152,10 +153,8 @@ class QualityPlugin implements Plugin<Project> {
                         configLoader.resolveCheckstyleConfig()
                         applyExcludes(it, extension)
                     }
-                    reports {
-                        xml.enabled = true
-                        html.enabled = extension.htmlReports
-                    }
+                    enableReport(reports.xml)
+                    enableReport(reports.html, extension.htmlReports)
                 }
             }
             configurePluginTasks(project, extension, Checkstyle, 'checkstyle', new CheckstyleReporter(configLoader))
@@ -191,10 +190,8 @@ class QualityPlugin implements Plugin<Project> {
                         configLoader.resolvePmdConfig()
                         applyExcludes(it, extension)
                     }
-                    reports {
-                        xml.enabled = true
-                        html.enabled = extension.htmlReports
-                    }
+                    enableReport(reports.xml)
+                    enableReport(reports.html, extension.htmlReports)
                 }
             }
             configurePluginTasks(project, extension, Pmd, 'pmd', new PmdReporter())
@@ -243,7 +240,7 @@ class QualityPlugin implements Plugin<Project> {
                     )))
                     reports {
                         xml {
-                            enabled true
+                            enableReport(it)
                         }
                     }
                 }
@@ -272,10 +269,8 @@ class QualityPlugin implements Plugin<Project> {
                         configLoader.resolveCodenarcConfig()
                         applyExcludes(it, extension)
                     }
-                    reports {
-                        xml.enabled = true
-                        html.enabled = extension.htmlReports
-                    }
+                    enableReport(reports.xml)
+                    enableReport(reports.html, extension.htmlReports)
                 }
             }
             configurePluginTasks(project, extension, CodeNarc, 'codenarc', new CodeNarcReporter())
@@ -348,9 +343,7 @@ class QualityPlugin implements Plugin<Project> {
             Class<Task> cpdTasksType = plugin.class.classLoader.loadClass('de.aaschmid.gradle.plugins.cpd.Cpd')
             // reports applied for all registered cpd tasks
             prj.tasks.withType(cpdTasksType).configureEach { task ->
-                reports {
-                    xml.enabled = true
-                }
+                enableReport(reports.xml)
                 doFirst {
                     configLoader.resolveCpdXsl()
                 }
@@ -367,6 +360,14 @@ class QualityPlugin implements Plugin<Project> {
             // yes, it's not completely normal that module could disable root project task, but it would be much
             // simpler to use like that (because quality plugin assumed to be applied in subprojects section)
             applyEnabledState(prj, extension, cpdTasksType)
+        }
+    }
+
+    private void enableReport(Report report, boolean enable = true) {
+        if (GradleVersion.current() < GradleVersion.version('7.0')) {
+            report.enabled = enable
+        } else {
+            report.required.set(enable)
         }
     }