+++ src/main/groovy/ru/vyarus/gradle/plugin/quality/QualityExtension.groovy
@@ -21,10 +21,10 @@ class QualityExtension {
         sourceSets = [project.sourceSets.main] as Collection<SourceSet>
     }
 
-    String checkstyleVersion = '8.42'
+    String checkstyleVersion = '9.1'
     String pmdVersion = '6.34.0'
-    String spotbugsVersion = '4.2.3'
-    String codenarcVersion = '2.1.0'
+    String spotbugsVersion = '4.4.2'
+    String codenarcVersion = '2.2.0'
 
     /**
      * Sets AnimalSniffer version.